revert Connection trait into a class
