+++ src/Database/Connections/MySqlConnection.php
@@ -7,13 +7,14 @@ use Illuminate\Database\Query\Processors\MySqlProcessor;
 use Winter\Storm\Database\PDO\MySqlDriver;
 use Winter\Storm\Database\Query\Grammars\MySqlGrammar as QueryGrammar;
 use Winter\Storm\Database\Schema\Grammars\MySqlGrammar as SchemaGrammar;
+use Winter\Storm\Database\Traits\HasConnection;
 
 /**
  * @phpstan-property \Illuminate\Database\Schema\Grammars\Grammar|null $schemaGrammar
  */
 class MySqlConnection extends \Illuminate\Database\MySqlConnection
 {
-    use Connection;
+    use HasConnection;
 
     /**
      * Get the default query grammar instance.
